Extract factories to reduce checks on null
--HG--
branch : 5.21.x

+++ DEV/org.openl.core/test/org/openl/meta/LongValueTest.java
@@ -85,7 +85,7 @@ public class LongValueTest {
     @Test
     public void testProduct() {
         LongValue[] la = getTestArray();
-        assertEquals(new DoubleValue(5000), LongValue.product(la));
+        assertEquals(new LongValue(5000), LongValue.product(la));
 
         LongValue[] nullArray = null;
         assertEquals(null, LongValue.product(nullArray));
